# Speed up fold ~3x