@acowley 2015-11-29
That's a pretty substantial speedup. Quick question: is it just the composition with `final` that's slowing it down, or is `scan` not getting inlined?