@YoEight 2015-11-29
@acowley IMHO, `(~>)` is at fault. That function isn't as fast as its `pipes` counterpart for instance.