@HuwCampbell 2015-11-30
I think it's that it's yielding a lot of intermediate values and discarding them, it's faster to just not yield them.