@acowley 2015-11-30
Yeah, that's the kind of thing that gets eaten up be RULES most of the time. We could probably do something specific for rewriting uses of `final` to avoid the machine composition. IIUC, [this](https://github.com/Gabriel439/Haskell-Pipes-Library/blob/master/src/Pipes/Core.hs#L308) is the relevant `pipes` correspondence. Is it actually doing anything terribly more efficient? (I honestly don't know)

The specific benchmark that @HuwCampbell is looking at is also kind of interesting as there's just not a lot to it. We should perhaps sprinkle INLINABLE pragmas more generously to crunch down on concrete types