@treeowl 2015-12-04
I don't see any reason not to accept this, although it seems to point to deeper efficiency concerns.